Update to Swift 4.2 + Linting
Update to Swift 4.2 and Linting

+++ ExtensionKit/ExtensionKit/UIKit/UIColor.swift
@@ -11,7 +11,7 @@ import Foundation
 public extension UIColor {
 	
 	public convenience init(hex: String, alpha: CGFloat = 1.0) {
-		let hex = UInt(hex: hex) ?? 0
+        let hex = UInt(hex: hex) ?? UInt(0)
 		self.init(red: CGFloat((hex & 0xFF0000) >> 16) / 255.0,
 		          green: CGFloat((hex & 0x00FF00) >> 8) / 255.0,
 		          blue: CGFloat(hex & 0x0000FF) / 255.0, alpha: alpha)
